+++ core/src/commonDom/kotlin/nl/adaptivity/xmlutil/util/CompactFragment.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021.
+ * Copyright (c) 2023.
  *
  * This file is part of xmlutil.
  *