+++ serialization/src/commonMain/kotlin/nl/adaptivity/xmlutil/serialization/XmlConfig.kt
@@ -377,6 +377,7 @@ private constructor(
                 typeDiscriminatorName = QName(XMLConstants.XSI_NS_URI, "type", XMLConstants.XSI_PREFIX)
                 encodeDefault = XmlEncodeDefault.ANNOTATED
                 throwOnRepeatedElement = true
+                isStrictAttributeNames = true
             }
         }